implemented sweetalert

+++ client/src/components/applicant-detail.ts
@@ -1,5 +1,6 @@
 import {inject} from 'aurelia-framework';
 import {EventAggregator} from 'aurelia-event-aggregator';
+import Swal from 'sweetalert2';
 import { ApplicantAPI } from './../api/agent';
 import {ApplicantUpdated,ApplicantViewed} from './messages';
 import {areEqual} from '../utility';
@@ -45,28 +46,71 @@ export class ApplicantDetail {
   }
 
   update(params) {
-    if (!areEqual(params, this.originalApplicant)) {
-      return this.api.update(params).then(newApplicant => {
-        this.newApplicant = <Applicant>newApplicant;
-        this.ea.publish(new ApplicantUpdated(this.newApplicant));
-      });
-      
-    }else{
-      
-      console.log(`There are no new changes`)
-    }
-    
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, update it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        if (!areEqual(params, this.originalApplicant)) {
+          return this.api.update(params).then(newApplicant => {
+            this.newApplicant = <Applicant>newApplicant;
+            this.ea.publish(new ApplicantUpdated(this.newApplicant));
+            Swal.fire(
+              'Updated!',
+              'Your record has been updated.',
+              'success'
+            )
+            setTimeout(function(){location.replace("/")}, 3000);
+          });
+        }else{
+          Swal.fire(
+            'Warning',
+            'There are no new changes',
+            'warning',
+          );
+        }
+        
+      }
+    })
   }
 
   delete(params) {
-    if (params !== undefined) {
-      return this.api.delete(params.id).then(responseMessage => {
-      });
-      
-    }else{
-      console.log(`The fields are empty.`)
-    }
-    
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        if (params !== undefined) {
+          return this.api.delete(params.id).then(responseMessage => {
+            Swal.fire(
+              'Deleted!',
+              'Your record has been deleted.',
+              'success'
+            )
+            setTimeout(function(){location.replace("/")}, 3000);
+          });
+          
+          
+        }else{
+          Swal.fire(
+            'Warning',
+            'The fields are empty!',
+            'warning',
+          );
+        }
+        
+      }
+    })
   }
 
   canDeactivate() {
